stop at tests

+++ tensorcircuit/gates.py
@@ -367,7 +367,7 @@ def meta_gate() -> None:
                 m = np.reshape(m, newshape=(2, 2, 2, 2))
             if m.shape[0] == 8:
                 m = np.reshape(m, newshape=(2, 2, 2, 2, 2, 2))
-            #m = m.astype(npdtype)
+            m = m.astype(npdtype)
             # not enough for new mechanism: register method on class instead of instance
             # temp = partial(gate_wrapper, m, n)
             # temp.__name__ = n
